+++ mvvm/src/main/java/com/catchpig/mvvm/network/factory/SerializationConverterFactory.kt
@@ -8,12 +8,12 @@ import retrofit2.Converter
 import retrofit2.Retrofit
 import java.lang.reflect.Type
 
-class SerializationCoverterFactory(private val serviceClassName: String) :
+class SerializationConverterFactory(private val serviceClassName: String) :
         Converter.Factory() {
     companion object {
         @JvmStatic
-        fun create(serviceClassName: String): SerializationCoverterFactory {
-            return SerializationCoverterFactory(serviceClassName)
+        fun create(serviceClassName: String): SerializationConverterFactory {
+            return SerializationConverterFactory(serviceClassName)
         }
     }